+++ dolfin/graph/SCOTCH.cpp
@@ -50,7 +50,7 @@ using namespace dolfin;
 void SCOTCH::compute_partition(
   const MPI_Comm mpi_comm,
   std::vector<std::size_t>& cell_partition,
-  std::map<std::size_t, dolfin::Set<unsigned int>>& ghost_procs,
+  std::map<std::int64_t, dolfin::Set<int>>& ghost_procs,
   const LocalMeshData& mesh_data)
 {
   // Create data structures to hold graph
@@ -68,8 +68,7 @@ void SCOTCH::compute_partition(
 
 }
 //-----------------------------------------------------------------------------
-std::vector<int> SCOTCH::compute_gps(const Graph& graph,
-                                     std::size_t num_passes)
+std::vector<int> SCOTCH::compute_gps(const Graph& graph, std::size_t num_passes)
 {
   // Create strategy string for Gibbs-Poole-Stockmeyer ordering
   std::string strategy = "g{pass= " + std::to_string(num_passes) + "}";
@@ -201,7 +200,7 @@ void SCOTCH::partition(
   const std::vector<std::size_t>& global_cell_indices,
   const std::size_t num_global_vertices,
   std::vector<std::size_t>& cell_partition,
-  std::map<std::size_t, dolfin::Set<unsigned int>>& ghost_procs)
+  std::map<std::int64_t, dolfin::Set<int>>& ghost_procs)
 {
   Timer timer("Compute graph partition (SCOTCH)");
 
@@ -290,9 +289,9 @@ void SCOTCH::partition(
   // Build SCOTCH distributed graph
   Timer timer1("SCOTCH: call SCOTCH_dgraphBuild");
   if (SCOTCH_dgraphBuild(&dgrafdat, baseval, vertlocnbr, vertlocnbr,
-                              &vertloctab[0], NULL, veloloctab, NULL,
-                              edgelocnbr, edgelocnbr,
-                              &edgeloctab[0], NULL, NULL) )
+                        &vertloctab[0], NULL, veloloctab, NULL,
+                        edgelocnbr, edgelocnbr,
+                        &edgeloctab[0], NULL, NULL) )
   {
     dolfin_error("SCOTCH.cpp",
                  "partition mesh using SCOTCH",
@@ -403,7 +402,7 @@ void SCOTCH::partition(
         auto map_it = ghost_procs.find(i);
         if (map_it == ghost_procs.end())
         {
-          dolfin::Set<unsigned int> sharing_processes;
+          dolfin::Set<int> sharing_processes;
           // Owning process goes first into dolfin::Set (unordered
           // set) so will always be first.
           sharing_processes.insert(proc_this);
@@ -430,19 +429,17 @@ void SCOTCH::partition(
 //-----------------------------------------------------------------------------
 #else
 //-----------------------------------------------------------------------------
-void SCOTCH::compute_partition(
-  const MPI_Comm mpi_comm,
-  std::vector<std::size_t>& cell_partition,
-  std::map<std::size_t, dolfin::Set<unsigned int>>& ghost_procs,
-  const LocalMeshData& mesh_data)
+void SCOTCH::compute_partition(const MPI_Comm mpi_comm,
+                               std::vector<std::size_t>& cell_partition,
+                               std::map<std::int64_t, dolfin::Set<int>>& ghost_procs,
+                               const LocalMeshData& mesh_data)
 {
   dolfin_error("SCOTCH.cpp",
                "partition mesh using SCOTCH",
                "DOLFIN has been configured without support for SCOTCH");
 }
 //-----------------------------------------------------------------------------
-std::vector<int> SCOTCH::compute_gps(const Graph& graph,
-                                     std::size_t num_passes)
+std::vector<int> SCOTCH::compute_gps(const Graph& graph, std::size_t num_passes)
 {
   dolfin_error("SCOTCH.cpp",
                "re-order graph using SCOTCH",
@@ -451,8 +448,7 @@ std::vector<int> SCOTCH::compute_gps(const Graph& graph,
 }
 //-----------------------------------------------------------------------------
 std::vector<int>
-SCOTCH::compute_reordering(const Graph& graph,
-                           std::string scotch_strategy)
+SCOTCH::compute_reordering(const Graph& graph, std::string scotch_strategy)
 {
   dolfin_error("SCOTCH.cpp",
                "re-order graph using SCOTCH",
@@ -478,7 +474,7 @@ void SCOTCH::partition(const MPI_Comm mpi_comm,
                        const std::vector<std::size_t>& global_cell_indices,
                        const std::size_t num_global_vertices,
                        std::vector<std::size_t>& cell_partition,
-                       std::map<std::size_t, dolfin::Set<unsigned int>>& ghost_procs)
+                       std::map<std::int64_t, dolfin::Set<int>>& ghost_procs)
 {
   dolfin_error("SCOTCH.cpp",
                "partition mesh using SCOTCH",
