+++ dolfin/graph/SCOTCH.h
@@ -22,6 +22,7 @@
 #define __SCOTCH_PARTITIONER_H
 
 #include <cstddef>
+#include <cstdint>
 #include <map>
 #include <set>
 #include <string>
@@ -50,7 +51,7 @@ namespace dolfin
     static void compute_partition(
       const MPI_Comm mpi_comm,
       std::vector<std::size_t>& cell_partition,
-      std::map<std::size_t, dolfin::Set<unsigned int> >& ghost_procs,
+      std::map<std::int64_t, dolfin::Set<int>>& ghost_procs,
       const LocalMeshData& mesh_data);
 
     /// Compute reordering (map[old] -> new) using
@@ -75,13 +76,13 @@ namespace dolfin
     // Compute cell partitions from distributed dual graph
     static void partition(
       const MPI_Comm mpi_comm,
-      const std::vector<std::set<std::size_t> >& local_graph,
+      const std::vector<std::set<std::size_t>>& local_graph,
       const std::vector<std::size_t>& node_weights,
       const std::set<std::size_t>& ghost_vertices,
       const std::vector<std::size_t>& global_cell_indices,
       const std::size_t num_global_vertices,
       std::vector<std::size_t>& cell_partition,
-      std::map<std::size_t, dolfin::Set<unsigned int> >& ghost_procs);
+      std::map<std::int64_t, dolfin::Set<int>>& ghost_procs);
 
   };
